+++ jobs/CHANGELOG.rst
@@ -11,3 +11,4 @@ This document describes changes between each past release.
 - Scrape product releases and nightlies from https://archives.mozilla.org
 - Fetch system addons update information from Balrog
 - Add mimetype support (#47)
+- Infer channel from URL (fixes #8)